suggestion: Refactor
