+++ cmd/suggestion/bayesianoptimization/README.md
@@ -1,0 +1,14 @@
+- start the service
+
+```
+python suggestion/bayesian/main.py
+```
+
+- start the testing client
+
+```
+python suggestion/test_client.py
+```
+
+note:
+the testing client uses the [Franke's function](http://www.sfu.ca/~ssurjano/franke2d.html) as the black box, and the maximum of Franke's function is around 1.22
