+++ cmd/suggestion/bayesianoptimization/requirements.txt
@@ -1,0 +1,4 @@
+grpcio
+numpy>=1.13.3
+scikit-learn>=0.19.0
+scipy>=0.19.1
