+++ cmd/suggestion/random/Dockerfile
@@ -1,0 +1,10 @@
+FROM golang:alpine AS build-env
+# The GOPATH in the image is /go.
+ADD . /go/src/github.com/kubeflow/katib
+WORKDIR /go/src/github.com/kubeflow/katib/cmd/suggestion/random
+RUN go build -o random
+
+FROM alpine:3.7
+WORKDIR /app
+COPY --from=build-env /go/src/github.com/kubeflow/katib/cmd/suggestion/random /app/
+ENTRYPOINT ["./random"]
