+++ pkg/suggestion/test_func.py
@@ -1,0 +1,9 @@
+import numpy as np
+
+
+def func(x1, x2):
+    return 0.75 * np.exp(-(9 * x1-2) ** 2 / 4 - (9 * x2-2) ** 2 / 4) + 0.75 * np.exp(
+        -(9 * x1 + 1) ** 2 / 49 - (9 * x2 + 1) / 10) + \
+           0.5 * np.exp(-(9 * x1 - 7) ** 2 / 4 - (9 * x2 - 3) ** 2 / 4) - 0.2 * np.exp(
+        -(9 * x1 - 4) ** 2 - (9 * x2 - 7) ** 2)
+
