+++ pkg/suggestion/types.py
@@ -1,0 +1,1 @@
+DEFAULT_PORT = "0.0.0.0:6789"
