--- suggestion/bayesianoptimization/Dockerfile
@@ -1,8 +1,0 @@
-FROM python:3
-
-ADD . /usr/src/app/github.com/kubeflow/katib
-WORKDIR /usr/src/app/github.com/kubeflow/katib/suggestion/bayesianoptimization
-RUN pip install --no-cache-dir -r requirements.txt
-ENV PYTHONPATH /usr/src/app/github.com/kubeflow/katib
-
-ENTRYPOINT ["python", "main.py"]
