--- suggestion/types.py
@@ -1,1 +1,0 @@
-DEFAULT_PORT = "0.0.0.0:6789"
